refactor: Controller 클래스 메서드 정리
