--- slipp/src/main/java/slipp/controller/ListUserController.java
@@ -1,25 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.tobe.view.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@Controller
-public class ListUserController {
-
-    // TODO 메서드가 디폴트면 실행 안 됨!
-    @RequestMapping(value = "/users", method = RequestMethod.GET)
-    public ModelAndView showUsers(HttpServletRequest req, HttpServletResponse resp) {
-        if (!UserSessionUtils.isLogined(req.getSession())) {
-            return new ModelAndView("redirect:/users/loginForm");
-        }
-
-        req.setAttribute("users", DataBase.findAll());
-        return new ModelAndView("/user/list.jsp");
-    }
-}
