+++ slipp/src/main/java/slipp/controller/LoginController.java
@@ -13,6 +13,11 @@ import javax.servlet.http.HttpSession;
 @nextstep.web.annotation.Controller
 public class LoginController {
 
+    @RequestMapping(value = "/users/loginForm", method = RequestMethod.GET)
+    public ModelAndView showLoginForm(HttpServletRequest req, HttpServletResponse resp) {
+        return new ModelAndView("/user/login.jsp");
+    }
+
     @RequestMapping(value = "/users/login", method = RequestMethod.POST)
     public ModelAndView showLogin(HttpServletRequest req, HttpServletResponse resp) {
         String userId = req.getParameter("userId");
@@ -30,5 +35,12 @@ public class LoginController {
             req.setAttribute("loginFailed", true);
             return new ModelAndView("/user/login.jsp");
         }
+    }
+
+    @RequestMapping(value = "/users/logout", method = RequestMethod.GET)
+    public ModelAndView showLogout(HttpServletRequest req, HttpServletResponse resp) {
+        HttpSession session = req.getSession();
+        session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
+        return new ModelAndView("redirect:/");
     }
 }
