--- slipp/src/main/java/slipp/controller/LogoutController.java
@@ -1,20 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.tobe.view.ModelAndView;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-@nextstep.web.annotation.Controller
-public class LogoutController {
-
-    @RequestMapping(value = "/users/logout", method = RequestMethod.GET)
-    public ModelAndView showLogout(HttpServletRequest req, HttpServletResponse resp) {
-        HttpSession session = req.getSession();
-        session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
-        return new ModelAndView("redirect:/");
-    }
-}
