--- slipp/src/main/java/slipp/controller/ProfileController.java
@@ -1,24 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.tobe.view.ModelAndView;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@nextstep.web.annotation.Controller
-public class ProfileController {
-    @RequestMapping(value = "/users/profile", method = RequestMethod.GET)
-    public ModelAndView showProfile(HttpServletRequest req, HttpServletResponse resp) {
-        String userId = req.getParameter("userId");
-        User user = DataBase.findUserById(userId);
-        if (user == null) {
-            throw new NullPointerException("사용자를 찾을 수 없습니다.");
-        }
-        req.setAttribute("user", user);
-        return new ModelAndView("/user/profile.jsp");
-    }
-}
