--- slipp/src/main/java/slipp/controller/UpdateFormUserController.java
@@ -1,26 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.tobe.view.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@Controller
-public class UpdateFormUserController {
-
-    @RequestMapping(value = "/users/updateForm", method = RequestMethod.GET)
-    public ModelAndView updateProfile(HttpServletRequest req, HttpServletResponse resp) {
-        String userId = req.getParameter("userId");
-        User user = DataBase.findUserById(userId);
-        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
-            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
-        }
-        req.setAttribute("user", user);
-        return new ModelAndView("/user/updateForm.jsp");
-    }
-}
