+++ slipp/src/main/java/slipp/controller/UserController.java
@@ -16,6 +16,11 @@ import javax.servlet.http.HttpServletResponse;
 public class UserController {
     private static final Logger log = LoggerFactory.getLogger(UserController.class);
 
+    @RequestMapping(value = "/users/form", method = RequestMethod.GET)
+    public ModelAndView showUserForm(HttpServletRequest req, HttpServletResponse resp) {
+        return new ModelAndView("/user/form.jsp");
+    }
+
     @RequestMapping(value = "/users/create", method = {RequestMethod.POST})
     public ModelAndView createUser(HttpServletRequest req, HttpServletResponse resp) {
         User user = new User(req.getParameter("userId"), req.getParameter("password"), req.getParameter("name"),
@@ -40,4 +45,35 @@ public class UserController {
         return new ModelAndView("redirect:/");
     }
 
+    @RequestMapping(value = "/users", method = RequestMethod.GET)
+    public ModelAndView showUsers(HttpServletRequest req, HttpServletResponse resp) {
+        if (!UserSessionUtils.isLogined(req.getSession())) {
+            return new ModelAndView("redirect:/users/loginForm");
+        }
+
+        req.setAttribute("users", DataBase.findAll());
+        return new ModelAndView("/user/list.jsp");
+    }
+
+    @RequestMapping(value = "/users/profile", method = RequestMethod.GET)
+    public ModelAndView showProfile(HttpServletRequest req, HttpServletResponse resp) {
+        String userId = req.getParameter("userId");
+        User user = DataBase.findUserById(userId);
+        if (user == null) {
+            throw new NullPointerException("사용자를 찾을 수 없습니다.");
+        }
+        req.setAttribute("user", user);
+        return new ModelAndView("/user/profile.jsp");
+    }
+
+    @RequestMapping(value = "/users/updateForm", method = RequestMethod.GET)
+    public ModelAndView updateProfile(HttpServletRequest req, HttpServletResponse resp) {
+        String userId = req.getParameter("userId");
+        User user = DataBase.findUserById(userId);
+        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
+            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
+        }
+        req.setAttribute("user", user);
+        return new ModelAndView("/user/updateForm.jsp");
+    }
 }
